refactor: organize components for tidy project
