+++ my-app/src/LogoutButton.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useAuth0 } from "@auth0/auth0-react";
 
-const LogoutButton = () => {
+export function LogoutButton() {
   const { logout } = useAuth0();
 
   return (
@@ -10,5 +10,3 @@ const LogoutButton = () => {
     </button>
   );
 };
-
-export default LogoutButton;
