+++ my-app/src/components/Nav/Profile.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { useAuth0 } from "@auth0/auth0-react";
-import {Avatar} from '@chakra-ui/react';
+import { Avatar } from '@chakra-ui/react';
 
-const Profile = () => {
+export function Profile() {
   const { user, isAuthenticated, isLoading } = useAuth0();
 
   if (isLoading) {
@@ -18,5 +18,3 @@ const Profile = () => {
     )
   );
 };
-
-export default Profile;
