+++ my-app/src/components/Nav/index.js
@@ -1,0 +1,1 @@
+export { Nav } from './Nav';
