+++ my-app/src/components/Search.js
@@ -1,14 +1,15 @@
 import React from "react"
 import { useParams } from "react-router-dom";
-import { useState, useEffect } from "react";
-import Feed from "./Feed";
 import { Button, Text, Flex, Center } from "@chakra-ui/react";
+import { Feed } from "./shared/Feed";
 
-export default function Search() {
+export function Search() {
     let params = useParams();
+
      function go(direction){
         window.location.href=(`/search/${params.regexp}/${params.amount}/${(parseInt(params.page) + direction)}`)
     }
+
     return <>
         <Flex>
             <Button onClick={()=>go(-1)}>⬅️</Button>
