+++ my-app/src/utils/dateUtils.js
@@ -1,0 +1,7 @@
+export function dt(ts){
+    if (ts && ts["$date"]){
+      return ts.$date.toString()
+    }else{
+      return new Date(ts).toString()
+    }
+}
